+++ src/index.ts
@@ -11,27 +11,6 @@ type AnyKeys<T> = { [P in keyof T]?: T[P] | any }
 type Without<T, U> = { [P in Exclude<keyof T, keyof U>]?: never }
 type XOR<T, U> = T | U extends object ? (Without<T, U> & U) | (Without<U, T> & T) : T | U
 
-/**
- * Specific region of endpoint.
- * @link https://docs.atlas.mongodb.com/api/data-api-resources/#regional-requests
- */
-export enum Region {
-  Virginia = 'us-east-1',
-  Oregon = 'us-west-2',
-  Ireland = 'eu-west-1',
-  Sydney = 'ap-southeast-2'
-}
-
-// https://docs.atlas.mongodb.com/api/data-api-resources/#base-url
-const getUrlEndpoint = (appId: string, region?: Region) => {
-  return region
-    ? `https://${region}.aws.data.mongodb-api.com/app/${appId}/endpoint/data/beta`
-    : `https://data.mongodb-api.com/app/${appId}/endpoint/data/beta`
-}
-const getActionUrl = (endpoint: string, action: string) => {
-  return `${endpoint}/action/${action}`
-}
-
 type ExtendBaseParams<T> = BaseParams & T
 interface BaseParams {
   dataSource?: string
@@ -43,24 +22,35 @@ interface BaseParams {
 interface BaseConfig {
   /**
    * Specific Data API key.
-   * @link https://docs.atlas.mongodb.com/api/data-api/#2.-create-a-data-api-key
+   * @link https://www.mongodb.com/docs/atlas/api/data-api/#2.-create-a-data-api-key
    */
   apiKey: string
 }
+
 interface UrlEndpointConfig extends BaseConfig {
   /**
    * Specific URL Endpoint.
-   * @link https://docs.atlas.mongodb.com/api/data-api/#3.-send-a-data-api-request
+   * @link https://www.mongodb.com/docs/atlas/api/data-api/#3.-send-a-data-api-request
    */
   urlEndpoint: string
 }
+
 interface PackEndpointConfig extends BaseConfig {
   /**
    * Specific Data App ID.
-   * @link https://docs.atlas.mongodb.com/api/data-api/#3.-send-a-data-api-request
+   * @link https://www.mongodb.com/docs/atlas/api/data-api/#3.-send-a-data-api-request
    */
   appId: string
-  region?: Region
+  /**
+   * Specific region name of endpoint.
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#regional-requests
+   */
+  region?: string
+  /**
+   * Specific cloud provider of endpoint.
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#regional-requests
+   */
+  cloud?: string
 }
 
 export type Config = XOR<UrlEndpointConfig, PackEndpointConfig>
@@ -116,10 +106,10 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Execute a API action.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/
    */
   public $$action<Result = unknown>(
-    name: string,
+    type: string,
     params: BaseParams = {},
     axiosConfig?: AxiosRequestConfig
   ): Promise<Result> {
@@ -132,14 +122,29 @@ export class MongoDBDataAPI<InnerDoc = Document> {
       return Promise.reject('Invalid params: dataSource, database, collection')
     }
 
+    // https://www.mongodb.com/docs/atlas/api/data-api-resources/#base-url
+    // https://www.mongodb.com/docs/atlas/api/data-api-resources/#regional-requests
+    const getUrlEndpoint = (appId: string, region?: string, cloud?: string) => {
+      return region && cloud
+        ? `https://${region}.${cloud}.data.mongodb-api.com/app/${appId}/endpoint/data/v1`
+        : `https://data.mongodb-api.com/app/${appId}/endpoint/data/v1`
+    }
+
+    const getActionUrl = (endpoint: string, action: string) => {
+      return `${endpoint}/action/${action}`
+    }
+
     const API_KEY_FIELD = 'api-key'
 
     return this.#axios({
       method: 'post',
       data: JSON.stringify(mergedParams),
       url: this.#config.urlEndpoint
-        ? getActionUrl(this.#config.urlEndpoint, name)
-        : getActionUrl(getUrlEndpoint(this.#config.appId!, this.#config.region), name),
+        ? getActionUrl(this.#config.urlEndpoint, type)
+        : getActionUrl(
+            getUrlEndpoint(this.#config.appId!, this.#config.region, this.#config.cloud),
+            type
+          ),
       headers: {
         'Content-Type': 'application/json',
         'Access-Control-Request-Headers': '*',
@@ -151,16 +156,20 @@ export class MongoDBDataAPI<InnerDoc = Document> {
         return response.data
       })
       .catch((error) => {
-        // https://docs.atlas.mongodb.com/api/data-api-resources/#error-codes
-        const errorJSON = error.toJSON()
-        errorJSON.config.headers[API_KEY_FIELD] = '*****'
-        return Promise.reject(error.toJSON())
+        // https://www.mongodb.com/docs/atlas/api/data-api-resources/#error-codes
+        if (_axios.isAxiosError(error)) {
+          const errorJSON: any = error.toJSON()
+          errorJSON.config.headers[API_KEY_FIELD] = '*****'
+          return Promise.reject(errorJSON)
+        } else {
+          return Promise.reject(error)
+        }
       })
   }
 
   /**
    * Find a Single Document.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#find-a-single-document
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#find-a-single-document
    */
   public findOne<D = InnerDoc, T = NoInfer<D>>(
     params?: ExtendBaseParams<{
@@ -173,7 +182,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Find Multiple Documents.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#find-multiple-documents
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#find-multiple-documents
    */
   public find<D = InnerDoc, T = NoInfer<D>>(
     params?: ExtendBaseParams<{
@@ -189,7 +198,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Insert a Single Document.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#insert-a-single-document
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#insert-a-single-document
    */
   public insertOne<D = InnerDoc, T = NoInfer<D>>(
     params: ExtendBaseParams<{ document: AnyKeys<T> | Document }>
@@ -199,7 +208,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Insert Multiple Documents.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#insert-multiple-documents
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#insert-multiple-documents
    */
   public insertMany<D = InnerDoc, T = NoInfer<D>>(
     params: ExtendBaseParams<{ documents: Array<AnyKeys<T> | Document> }>
@@ -209,7 +218,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Update a Single Document.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#update-a-single-document
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#update-a-single-document
    */
   public updateOne<D = InnerDoc, T = NoInfer<D>>(
     params: ExtendBaseParams<{
@@ -227,7 +236,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Update Multiple Documents.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#update-multiple-documents
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#update-multiple-documents
    */
   public updateMany<D = InnerDoc, T = NoInfer<D>>(
     params: ExtendBaseParams<{
@@ -245,7 +254,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Replace a Single Document.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#replace-a-single-document
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#replace-a-single-document
    */
   public replaceOne<D = InnerDoc, T = NoInfer<D>>(
     params: ExtendBaseParams<{
@@ -263,7 +272,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Delete a Single Document.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#delete-a-single-document
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#delete-a-single-document
    */
   public deleteOne<D = InnerDoc, T = NoInfer<D>>(
     params: ExtendBaseParams<{ filter: Filter<T> }>
@@ -273,7 +282,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Delete Multiple Documents.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#delete-multiple-documents
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#delete-multiple-documents
    */
   public deleteMany<D = InnerDoc, T = NoInfer<D>>(
     params: ExtendBaseParams<{ filter: Filter<T> }>
@@ -283,7 +292,7 @@ export class MongoDBDataAPI<InnerDoc = Document> {
 
   /**
    * Run an Aggregation Pipeline.
-   * @link https://docs.atlas.mongodb.com/api/data-api-resources/#run-an-aggregation-pipeline
+   * @link https://www.mongodb.com/docs/atlas/api/data-api-resources/#run-an-aggregation-pipeline
    */
   public aggregate<T extends Array<any>>(
     params: ExtendBaseParams<{ pipeline: Array<Document> }>